+++ src/game/resource_manager.cpp
@@ -22,12 +22,12 @@ struct StbiImage
     StbiImage(const std::string& path, TextureInfo& info) : buffer(NULL)
     {
         SDL_RWops* f = SDL_RWFromFile(path.c_str(), "rb");
-        i64 size = SDL_RWsize(f);
+        i32 size = (i32)SDL_RWsize(f);
         u8* data = new u8[size];
         SDL_RWread(f, data, size, 1);
         SDL_RWclose(f);
 
-        buffer = stbi_load_from_memory(data, (int)size, &info.Width, &info.Height, &info.ComponentCount, 0);
+        buffer = stbi_load_from_memory(data, size, &info.Width, &info.Height, &info.ComponentCount, 0);
         if (buffer == NULL)
             printf("failed to load image: %s", stbi_failure_reason());
